# docs/api-docs.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "ICoder API",
    "description": "API Documentation for ICoder Project",
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "http://localhost:9090",
      "description": "Generated server url"
    }
  ],
  "paths": {
...
    "/api/v1/activity-streak": {
      "get": {
        "tags": [
          "activity-streak"
        ],
        "summary": "Get User Streak",
        "description": "Retrieve the current streak of the user",
        "operationId": "getUserStreak",
        "parameters": [
          {
            "name": "timezone",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "default": "UTC"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/StreakResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/activity-logs/my": {
      "get": {
        "tags": [
          "Activity Logs"
        ],
        "summary": "Get My Activity Logs",
        "description": "Retrieve a paginated list of activity logs for the current user",
        "operationId": "getMyActivityLogs",
        "parameters": [
          {
            "name": "pageable",
            "in": "query",
            "required": true,
            "schema": {
              "$ref": "#/components/schemas/Pageable"
            }
          },
          {
            "name": "verdict",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "enum": [
                "PENDING",
                "ACCEPTED",
                "WRONG_ANSWER",
                "IN_QUEUE",
                "COMPILATION_ERROR",
                "RUNTIME_ERROR",
                "TIME_LIMIT_EXCEEDED",
                "FAILED",
                "RUNNING",
                "MEMORY_LIMIT_EXCEEDED",
                "FAILED_TO_SUBMIT"
              ]
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/PageActivityLogResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/activity-logs/grid": {
      "get": {
        "tags": [
          "Activity Logs"
        ],
        "summary": "Get Activity Grid",
        "description": "Retrieve a list of activity grid data for the current user",
        "operationId": "getActivityGrid",
        "parameters": [
          {
            "name": "year",
            "in": "query",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          },
          {
            "name": "timezone",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "default": "UTC"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ActivityGridResponse"
                  }
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "Pageable": {
        "type": "object",
        "properties": {
          "page": {
            "minimum": 0,
            "type": "integer",
            "format": "int32"
          },
          "size": {
            "minimum": 1,
            "type": "integer",
            "format": "int32"
          },
          "sort": {
            "type": "array",
            "items": {
              "type": "string"
            }
          }
        }
      },
...
      "PageableObject": {
        "type": "object",
        "properties": {
          "paged": {
            "type": "boolean"
          },
          "pageNumber": {
            "type": "integer",
            "format": "int32"
          },
          "pageSize": {
            "type": "integer",
            "format": "int32"
          },
          "unpaged": {
            "type": "boolean"
          },
          "offset": {
            "type": "integer",
            "format": "int64"
          },
          "sort": {
            "$ref": "#/components/schemas/SortObject"
          }
        }
      },
      "SortObject": {
        "type": "object",
        "properties": {
          "sorted": {
            "type": "boolean"
          },
          "unsorted": {
            "type": "boolean"
          },
          "empty": {
            "type": "boolean"
          }
        }
      },
...
      "StreakResponse": {
        "type": "object",
        "properties": {
          "current_streak": {
            "type": "integer",
            "format": "int32"
          },
          "max_streak": {
            "type": "integer",
            "format": "int32"
          },
          "last_accepted_at": {
            "type": "string",
            "format": "date-time"
          },
          "today_utc": {
            "type": "string"
          }
        }
      },
      "ActivityLogResponse": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "user_handle": {
            "type": "string"
          },
          "activity_type": {
            "type": "string",
            "enum": [
              "SUBMISSION",
              "JOIN",
              "SOLVED_PROBLEM"
            ]
          },
          "entity_type": {
            "type": "string"
          },
          "verdict": {
            "type": "string",
            "enum": [
              "PENDING",
              "ACCEPTED",
              "WRONG_ANSWER",
              "IN_QUEUE",
              "COMPILATION_ERROR",
              "RUNTIME_ERROR",
              "TIME_LIMIT_EXCEEDED",
              "FAILED",
              "RUNNING",
              "MEMORY_LIMIT_EXCEEDED",
              "FAILED_TO_SUBMIT"
            ]
          },
          "entity_id": {
            "type": "integer",
            "format": "int64"
          },
          "created_at": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
      "PageActivityLogResponse": {
        "type": "object",
        "properties": {
          "totalElements": {
            "type": "integer",
            "format": "int64"
          },
          "totalPages": {
            "type": "integer",
            "format": "int32"
          },
          "pageable": {
            "$ref": "#/components/schemas/PageableObject"
          },
          "first": {
            "type": "boolean"
          },
          "last": {
            "type": "boolean"
          },
          "size": {
            "type": "integer",
            "format": "int32"
          },
          "content": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ActivityLogResponse"
            }
          },
          "number": {
            "type": "integer",
            "format": "int32"
          },
          "sort": {
            "$ref": "#/components/schemas/SortObject"
          },
          "numberOfElements": {
            "type": "integer",
            "format": "int32"
          },
          "empty": {
            "type": "boolean"
          }
        }
      },
      "ActivityGridResponse": {
        "type": "object",
        "properties": {
          "date": {
            "type": "string",
            "format": "date"
          },
          "accepted_count": {
            "type": "integer",
            "format": "int64"
          },
          "attempted_count": {
            "type": "integer",
            "format": "int64"
          }
        }
      }
    },
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  }
}